requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
boto3==1.28.35
botocore==1.31.35
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.4
django-admin-sortable2==2.1.9
django-appconf==1.0.5
django-colorfield==0.10.1
django-embed-video==1.4.9
django-environ==0.10.0
django-heroku==0.3.1
django-image-optimizer==1.0.3
django-imagekit==4.1.0
django-ranged-response==0.2.0
django-s3-storage==0.14.0
django-simple-captcha==0.5.18
django-taggit==4.0.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
jmespath==1.0.1
Markdown==3.4.4
packaging==23.1
pilkit==2.0
Pillow==9.5.0
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyembed==1.3.3
pyembed-markdown==1.1.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-resize-image==1.1.20
requests==2.31.0
s3transfer==0.6.2
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tinify==1.6.0
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.7.1
urllib3==1.26.16
whitenoise==6.5.0
zipp==3.16.2
zope.interface==6.0